WIP volumes

+++ kohi/code/kohi.core.tests/src/containers/array_tests.c
@@ -69,11 +69,14 @@ static u8 array_all_iterator_tests(void) {
             u8* val = it.value(&it);
             if (it.pos == 0) {
                 expect_should_be(69, *val);
-            } else if (it.pos == 2) {
+            }
+            else if (it.pos == 2) {
                 expect_should_be(42, *val);
-            } else if (it.pos == 4) {
+            }
+            else if (it.pos == 4) {
                 expect_should_be(36, *val);
-            } else {
+            }
+            else {
                 expect_should_be(0, *val);
             }
 
@@ -84,18 +87,21 @@ static u8 array_all_iterator_tests(void) {
         // Try reverse/backward iteration.
         it = arr.rbegin(&arr.base);
         expect_should_be(&arr.base, it.arr);
-        expect_should_be(arr.base.length - 1, it.pos);
+        expect_should_be(arr.base.length - 1, (u32)it.pos);
         expect_should_be(-1, it.dir);
         loop_count = 0;
         for (; !it.end(&it); it.next(&it)) {
             u8* val = it.value(&it);
             if (it.pos == 0) {
                 expect_should_be(69, *val);
-            } else if (it.pos == 2) {
+            }
+            else if (it.pos == 2) {
                 expect_should_be(42, *val);
-            } else if (it.pos == 4) {
+            }
+            else if (it.pos == 4) {
                 expect_should_be(36, *val);
-            } else {
+            }
+            else {
                 expect_should_be(0, *val);
             }
